extract class out to its own class
