+++ app/src/main/java/com/cmput301w23t09/qrhunter/qrcode/QRCodeFragment.java
@@ -199,7 +199,7 @@ public class QRCodeFragment extends DialogFragment implements Serializable {
                                   .reduce(0, Integer::sum);
 
                           // Add an entry into our adapter with their score and player.
-                          scansAdapter.add(new QRCodePlayerScansAdapter.Entry(player, score));
+                          scansAdapter.add(new QRPlayerScanEntry(player, score));
                         });
               });
     }
